--- src/view/GamesContent.tsx
@@ -1,23 +1,0 @@
-import { Alert, AlertIcon, Stack } from '@chakra-ui/react'
-import { useContext } from 'react'
-import authContext from '../context/auth'
-import { GamesReader } from '../reader/game'
-import GameItemView from './GameItem'
-
-export default function GamesContentView (): JSX.Element {
-  const authState = useContext(authContext)
-  const authed = authState.authed === true
-  if (!authed) {
-    return (
-      <Alert status='warning'>
-        <AlertIcon />
-        You are not signed in yet.
-      </Alert>
-    )
-  }
-  return (
-    <Stack>
-      <GamesReader DocView={GameItemView} />
-    </Stack>
-  )
-}
