+++ src/view/Hand.tsx
@@ -4,7 +4,7 @@ import { playerContext } from '../reader/player'
 
 export default function HandView (): JSX.Element {
   const playerState = useContext(playerContext)
-  const hand = playerState.hand?.map((rank) => <Text key={rank}>{rank}</Text>)
+  const hand = playerState.hand?.map((rank, index) => <Text key={index}>{rank}</Text>)
   return (
     <>
       <Text>Hand:</Text>
