+++ src/view/Header.tsx
@@ -8,12 +8,13 @@ import Curtain from './Curtain'
 export default function HeaderView (): JSX.Element {
   const authState = useContext(authContext)
   const authed = authState.authed === true
-  const gamesLinkView = <ChakraLinkView to='/games'>Games</ChakraLinkView>
   return (
     <HStack spacing='20px'>
       <AuthView />
       <ChakraLinkView to='/'>Home</ChakraLinkView>
-      <Curtain open={authed} openElement={gamesLinkView} />
+      <Curtain open={authed}>
+        <ChakraLinkView to='/games'>Games</ChakraLinkView>
+      </Curtain>
     </HStack>
   )
 }
