--- src/view/writer/JoinGame.tsx
@@ -1,11 +1,0 @@
-import Writer from '.'
-
-export default function JoinGameWriter ({ gameId }: { gameId: string }): JSX.Element {
-  return (
-    <Writer
-      fn='joinGame'
-      label='Join Game'
-      props={{ gameId }}
-    />
-  )
-}
